fix visualizer crashes on some unsupported display type questions
