+++ frontend/src/metabase/visualizations/index.ts
@@ -11,10 +11,7 @@ import type {
 } from "metabase-types/api";
 
 import type { RemappingHydratedDatasetColumn } from "./types";
-import type {
-  Visualization,
-  VisualizationSettingDefinition,
-} from "./types/visualization";
+import type { Visualization } from "./types/visualization";
 
 const visualizations = new Map<VisualizationDisplay, Visualization>();
 const aliases = new Map<string, Visualization>();
@@ -157,23 +154,6 @@ export function isCartesianChart(display: VisualizationDisplay) {
     settingNames.includes("graph.dimensions") &&
     settingNames.includes("graph.metrics")
   );
-}
-
-const isDataSetting = ({
-  widget,
-}: VisualizationSettingDefinition<unknown, unknown>) => {
-  // TODO Come up with a better condition
-  return widget === "field" || widget === "fields";
-};
-
-export function getColumnVizSettings(display: VisualizationDisplay) {
-  const visualization = visualizations.get(display);
-  const settings = visualization?.settings ?? {};
-
-  // TODO Come up with a better condition
-  return Object.keys(settings).filter((key) => {
-    return isDataSetting(settings[key] ?? {});
-  });
 }
 
 // removes columns with `remapped_from` property and adds a `remapping` to the appropriate column
